Move balance endpoint from user controller to payment controller and remove user controller
