+++ packages/apps/job-launcher/server/src/common/constants/errors.ts
@@ -56,7 +56,6 @@ export enum ErrorEscrow {
 export enum ErrorUser {
   NotFound = 'User not found.',
   AccountCannotBeRegistered = 'Account cannot be registered.',
-  BalanceCouldNotBeRetrieved = 'User balance could not be retrieved.',
   InvalidCredentials = 'Invalid credentials.',
   UserNotActive = 'User not active.',
   DuplicatedEmail = 'The email you are trying to use already exists. Please check that the email is correct or use a different email.',
@@ -108,6 +107,7 @@ export enum ErrorPayment {
   InvalidRecipient = 'Invalid recipient',
   ChainIdMissing = 'ChainId is missing',
   InvalidChainId = 'Invalid chain id',
+  BalanceCouldNotBeRetrieved = 'User balance could not be retrieved.',
 }
 
 /**
